% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_template.R
\name{report_template}
\alias{report_template}
\title{Generate Latex Report Template.}
\usage{
report_template()
}
\description{
Generate Latex Report Template.
}
\details{
Creates a folder under \code{getwd()} named \code{report_Sys.Date()}. Folder contains file \code{report.Rnw} which is automatically opened in the editor, as well as \code{references.bib} for bibtex references.
}
\examples{
# setwd(...path...)
# report_template()
}
